Updated openstack/openstack Project: openstack/heat  592346eacbf0448b31d50e5ab6702bd17039525e
Only reset current resource's attrs in state_set

There is no need to reset other resources' attributes in the stack
when a resource calls state_set(), the other resources attributes will
get updated as they are visited. This is not a huge performance gain
since those resources would have been in memory anyway, but let's not
use the big hammer if we don't have to.

Change-Id: Icafd770055a3dc2f1f3262753a8dca045e38972d