Updated openstack/openstack Project: openstack-infra/project-config  788bb1e5229312b1cd91ee0c35ec9eaa2a2a81de
Enable NOVA_USE_SERVICE_TOKEN in nova-next job for Pike

We want to start testing a feature added to nova which
sends a service user token along with the regular user
token to other services like cinder and neutron during
long-running operations. This relies on a change in
devstack and we only run it on branches after stable/ocata.

Change-Id: Iee859b3a09066dcbc60583210b4107b787f0d79a
Depends-On: I2d7348c4a72af96c0ed2ef6c0ab75d16e9aec8fc