Updated openstack/openstack Project: openstack/nova  221bc57dee073e4c9d48611ff6eb0ae25e2a9b7d
docs: sort the Architecture Concepts index

The unsorted nature of this has always annoyed me, and since
I'm going to add a new item to the list in a following change
let's sort the list. Note that the list is sorted by title
in the linked doc, not the name of the doc file itself.

Change-Id: I24671fe5256889c3abc1ba36d70a450213f20234