Updated openstack/openstack Project: openstack/nova  a9562bcd6b1ccf5ab1655d3889e0a15852b762d5
Make scheduler get hosts from all cells

The scheduler needs to list all compute nodes and services in order to
have a list to filter. It does that based on information that it gets
from placement, but it needs to fetch the records out of the database.
This patch makes it do that for all cells, and takes the same approach
that compute api currently does, which is a refreshed-once-per-run
master list of cells that we can optimize later.

For the moment, the scheduler maintains a static list of cells, just
like compute/api does. Ideally before we release pike we will have
a better plan for making this dynamic.

Change-Id: I1a9b054b7ddace1b342ec82a86cb151d45ec8ae2