Updated openstack/openstack Project: openstack-dev/grenade  93fbeab01c58cbcb4a5499dc085f417dd72b84f3
Use dsconf to add lines to devstack localrc

In some situations there are additional sections in the devstack
localrc file and they may be at the end of the file. This caused
errors as the desired changes would not be placed in the
'[[local|localrc]]' section of the file.

Using 'dsconf setlc_raw' solves that issue.

Change-Id: If4bc197d685e20fadbc5ffbd2151cae9d8129135