Updated openstack/openstack Project: openstack/nova-specs  4ff806fb7c00baa51d3dd81716c3d0357c9ee28f
Enable cold migration with target host

The target host can be specified when cold migrating a server.
And a new request body field called ``force`` (defaulted to False)
is added in order not to call the scheduler when providing a
destination. Setting it to 'True' makes the cold migrate operation
bypass the scheduler.

Previously-approved: ocata

APIImpact
Change-Id: I610da96ebe8db892652e7c17d0606066bf604638
blueprint cold-migration-with-target-pike