Updated openstack/openstack Project: openstack/heat  b2421181fb31eccb1e2b6b79e1dbaeb7f7cab89a
Revert "Using os-testr to get better report"

This reverts commit d7f2d4f23eba13eea6c314bf5fde34df7a4df912.

ostestr spews what is IMHO an unmanageable amount of mostly worthless data when running tests - at least one line per test, of which there are 6351 at last count. There doesn't seem to be an option to quiet it down, so just revert it.

Change-Id: I7b12dc249e3e7041ab6f0ef36b3e5f0b082aeb7d