Updated openstack/openstack Project: openstack/heat  e26f0948535c140f048e5a2681576afe894a486f
Remove unused code in prepare_ports_for_replace()

At the beginning, when we prepare ports for
server replacement, we store 'fixed_ips' of
server ports before update them to []. But
update 'fixed_ips' is incorrect, we just have to
detach the ports from server. we fixed it already,
but seems the code was not cleaned.

Change-Id: If925e5bf03385922ec1784ae56ec7faff7c68b21