Updated openstack/openstack Project: openstack/swift  40fefd20870c1f812fea9f6f6d38dce2b807dba4
Probe tests fail, as requests checks for strings.

The requests library checks that the headers are either strings or
bytes. Currently, the two test_object_expirer tests fail with the
message:
InvalidHeader: Header value 1487879553 must be of type str
or bytes, not <type 'int'>

The header in question is "x-delete-at". The patch converts it to
a string, before making a Swift Client request.

Change-Id: I738697cb6b696f0e346345f75e0069048961f2ff