Updated openstack/openstack Project: openstack-dev/devstack  947af3ac616654aa39175ed9a3091454ebc4b525
Removed neutron_plugin_configure_debug_command functions

Those are not called by devstack anymore. This cleanup also gets rid of
code that attempts to set external_network_bridge to an empty value,
which triggers a deprecation warning for the option since it's going to
be removed in a next Neutron release.

Change-Id: I5adcbab877b4e8742522de81b1a85acfc33160d7