Updated openstack/openstack Project: openstack-infra/project-config  15fa2e7502af778da654f9bda0e72b8486ff109c
add tripleo non-voting job into tacker gate

With the motivation of adding more diversity in how do we test
tacker, this patch will add a tripleo job, non-voting.

The idea is to bring more feedback when testing patches in these
projects and detect what failures we could have outside devstack.

The jobs are not voting. We don't expect tacker team to open TripleO job
logs. The idea is to let TripleO folks know when a job is red so they
can investigate and tell if something critical is broken in the patch.

Change-Id: If2639be47e61649a220324e32d91dd9690a4f7e9