Updated openstack/openstack Project: openstack-infra/elastic-recheck  9d2398216dcbf35c34e77de9346acccde60c5cfe
Remove query for bug 1323658

This query is overly broad and is matching all (or many) the ssh or
networking failures from scenario tests. But, not specifically for the
bug listed, which was for a specific case where nova resize or restart
would break networking. Lets just remove the query and we can open a new
one for a bug that's specifically about the networking failure.

Change-Id: Iece853bc1a4c35a4d8c87a12665b0fcbed640776