Updated openstack/openstack Project: openstack/tempest  62424cb006c433dd14e32fc8755d2222276f95e1
Use base.create_flavor in servers test

This is to replace the multi-lines of creating flavor with
one call to base.create_flavor, so the code will look simpler.

Besides, _get_unused_flavor_id in ServersAdminNegativeTestJSON is not
necessary, for the random flavor id can hardly duplicate with the
existing ones.

Change-Id: I752b71860560a1a6c9368d07578a09b964859067