Updated openstack/openstack Project: openstack-infra/project-config  7dad5ebbdf82e886c86a77b4e5b02519a319d759
KEEP_LOCALRC for barbican-tempest-plugin

With the recent modifications to merge local.conf and localrc,
setting local.conf values in the pre_test_hook.sh get overwritten
unless KEEP_LOCALRC is set to 1.

It is understood that setting configuration options via
DEVSTACK_LOCAL_CONFIG is preferable, but it is beneficial to
be able to test configuration options alongside the change
being proposed.

Change-Id: Ib06ae8aa890d09ac2cd9e3bf4bfd45440618dcc2