Updated openstack/openstack Project: openstack-dev/pbr  028f82e6a7584127ae4590be1b0d42e50d60e539
doc: Clarify sections in 'setup.cfg'

Clarify what tools provide what sections in 'setup.cfg', thus explaining
why, for example, I couldn't find any references to '[build_sphinx]' in
either the pbr or setuptools source.

Comments are not a section, so this little bit of info is moved to a
'note'.

Change-Id: Icfb58195c58813e98ab48943119f53c7711331ec