Updated openstack/openstack Project: openstack/nova  3f5b5c9a31276fa7738287703107558c72ff0267
api/metadata/vendordata_dynamic: don't import Requests for its constants

Python-requests isn't really needed, don't import it to prevent
copy/paste of the anti pattern of accessing it's codes directly.

Fix one unit test that was returning a 204 but actually including data
in it, but expecting that data to not be used.

Change-Id: I9f6eda0c3e6f7d390abca5e9e4ebe059788ee11f