Updated openstack/openstack Project: openstack/nova  9d1e813e61a202f0e6caa59cc3015265f2355986
Ignore deleted services in minimum version calculation

When we go to detect the minimum version for a given service, we
should ignore any deleted services. Without this, we will return
the minimum version of all records, including those that have been
deleted with "nova service-delete". This patch filters deleted
services from the query.

Closes-Bug: #1668310
Change-Id: Ic96a5eb3728f97a3c35d2c5121e6fdcd4fd1c70b