Updated openstack/openstack Project: openstack/manila  a966e43b04f6bc8345f5939b36fb7064ddf3a290
[Grenade] Fix devstack configuration in CI hook

Commit [1] fixed approach for setting devstack config according to
changes [2] in devstack. But it started updating only 1 of 2 possible
devstack directories that are 'old' and 'new'. Where 'old' is the one
that is used only in 'grenade' job as base installation for performing
project migration. 'new' is used in all other single-devstack
installations. So, with [1] everything started working except
'grenade' job.
Fix grenade job by copying devstack config to 'old' dir if it exists.

[1] Idfa8e0cb3f446361daa1449e0854d95cfb379f46
[2] http://lists.openstack.org/pipermail/openstack-dev/2017-February/112872.html

Change-Id: Ic00a1fe296fa68a65672fa680ebd8a9336ed7cef