Updated openstack/openstack Project: openstack/neutron-lbaas  63f9d96f9814c5e3a7a01a6c52bdd8b72f9150bd
Work around devstack selecting incorrect image

Currently devstack is selecting the amphora image instead of the
cirros image to boot our test webservers.  This is a temporary fix
to get our gates functional again.  This will be removed when we can
come up with a better solution or devstack merges their fix [1].

[1] https://review.openstack.org/#/c/435106

Change-Id: Ia134c39277f30d1183c7585b09328037ded147f8