Updated openstack/openstack Project: openstack-infra/tripleo-ci  3aaf3fb2a559d39679dc2fd0d1fd2dd108b5c9a9
Update pip in DLRN venv

Newer versions of setuptools require an updated pip to be installed.
Running the --delorean-setup locally is actually failing for me with
(not sure what the difference is compared to CI, but probably is due to
using the infra created CentOS image which has updated pip/virtualenv
already):

ImportError: No module named six.moves

This is an expected problem with newer versions of setuptools per
https://github.com/pypa/setuptools/issues/942#issuecomment-274817619
https://github.com/pypa/setuptools/issues/937

We're already using the latest packaged versions of virtualenv and pip
on CentOS so the fix is to first update pip in the newly created
virtualenv.

Change-Id: Ibc0e266d48dec587c8221f8392f060b58496def4