Updated openstack/openstack Project: openstack/nova  30cb671bda7198500a7e642df4d3b284837993d5
Remove invalid tests-py3 whitelist item

The tests for XenAPI plugins were introduced with change [1].
This change also appended them to the python3 test whitelist,
as the functionality could only be executed on python2.4 environment.
The aforementioned changes were removed with change [2], but
the whitelist item was left behind. This change removes the obsolete
line.

[1] 2aa81e9ef6d04f2fdf1c69e16ceeef156cefedfa
[2] 21149582ff944585fb0a815ebe9515c1ec38eb7f

Change-Id: Ibc15e9a06adf34512eda904d0bb3eba1eed04784