Updated openstack/openstack Project: openstack-infra/project-config  d6c147bfa3426d6c6b93a160fd828f0bd057b22b
Detect misaligned ':' in zuul/layout.yaml

This came up in I79ced2532d03701f50afc9d2a6e4a6e533e8228a.  I don't
think it's actually a problem as the value is stripped.  Certainly the
zuul config-parser doesn't trip up on it.  So it's just for
consistency.

This found the entries fixed up in in zuul/layout.yaml

Change-Id: I76abb096426e24683fe8316c2f8e2d7a796dd96e