Updated openstack/openstack Project: openstack/neutron  a8c5bf0031b6c8851a889a69c9dfb6363dd9ff3b
Prevent double-subscribes with registry decorator

This adds a simple flag to object instances to avoid multiple calls
to 'subscribe' for a class with multiple children that use the
'has_registry_receivers' decorator.

The registry manager protects from multiple subscriptions of the
same function by storing callbacks based on IDs, but this is necessary
to eliminate the duplicate debug logging statements.

Change-Id: I38e0a5f4c361435ae4af6c16a407191b1e722e37