Updated openstack/openstack Project: openstack/cinder  923d5c34806b4ee83d9234233e3e32d08d0fc755
Reuse identical API v2 code for v1

There are a couple modules that are virtually identical between
API v1 and v2. This cleans up some of the v1 API code to just
use the v2 API code.

More clean up can be done for other API modules, but there are
subtle differences (returning 200 vs 202) between the two. The
code can still be reused, but some v1 specific handling will
need to be put in place to make that optimization.

Change-Id: Ice3b2819b65c55cb189a0c16c0d7ef2795bd20dd
Partial-bug: #1627921