Updated openstack/openstack Project: openstack/tripleo-heat-templates  6a8850e10ed892a2dedd1af02425fc3ffc43fc5b
Write out a json file containing container startup info and create tool to use it.

This adds a bit to the post.yaml for docker to write out a json file
containing all the information on how we are start docker containers
(thanks Dan!).  I've then written a script that parses this that can
be used to execute docker run commands in various ways for debugging
purposes.

Change-Id: I36d66b42d1ac5030db8841820d4fc512a71d1285
Co-Authored-by: Dan Prince <dprince@redhat.com>