Updated openstack/openstack Project: openstack/cinder  a72eae82cc20b842e4b57eb46e3be557269bc208
Change volume_type dict to ovo

The volume type in the create method in volume API
was changed from dict to ovo by the following patch:
  https://review.openstack.org/#/c/406780/
However, this is not changed in create group from src
API which calls volume create and therefore introduced
this bug. It throws an exception when calling volume_type.id
because volume_type is still a dict.

This patch fixed this problem.

Change-Id: I63cb785d27fa9e43da16a27da6d7b92052badf06
Closes-Bug: #1665549