Updated openstack/openstack Project: openstack-infra/devstack-gate  8b2ee25eeb37e3de1780c59056a1e9307d48ef08
Support setting a variable to get local.conf merged late

The current mechanisms for changing local.conf make it so that they
all happen before jobs get running. However projects like neutron do
some rather complicated per job configuration that has to happen on
top of devstack-gate configs. So expose a way for this to merge late.

This does remain a bit error prone because you are unaware of what is
being configured in devstack-gate first. It might be better to make
local.conf configuration either d-g or project oriented.

Change-Id: I53119876b82eeaec19382dbe115cb1740bed0cec