Updated openstack/openstack Project: openstack/nova-specs  3b7419cc2ceba4f9f4a90a2faf92a2c9e3d628a7
Add prep-for-network-aware-scheduling-pike spec

Change how we talk to neutron to allow us in the future to implement
network aware scheduling.

This is continuation of the work started during newton.

This is a important, because this blocks routed networks, which is
blocking the removal of cells v1.

blueprint prep-for-network-aware-scheduling-pike
Previously-Approved: Ocata

Change-Id: I4d87fd0cc8148aa90a1968064378b11a0d7b65c1