Updated openstack/openstack Project: openstack/glance  a36535787382fc347c98c71b099d6de5ac367faa
Mock CURRENT_RELEASE for migration unit test

The test_migrate unit test assumes that the CURRENT_VERSION set
in migrations.py is 'ocata'.  Thus this test will break when
the alembic constants are updated for the Pike release.  This
patch fixes the test by mocking the CURRENT_VERSION constant
with an appropriate value.

Change-Id: I2a7dc86b88afaf0e5b878d79607f54125a35eb16