Updated openstack/openstack Project: openstack/docs-specs  4d20a0afd4637bb637fde7922566d11b096dbfe6
Move administrator guide to project repos

The number of approved OpenStack projects continues to grow while the
documentation team resources are shrinking. We need a new approach to maintain
the high quality of the OpenStack documentation set.

This specification proposes allowing project teams to manage their
administrator guide content similar to how the api-ref and installation guides
are being managed.

Change-Id: I70592f5c8a415dc1454a89f1012ad50178359953