Updated openstack/openstack Project: openstack/keystone-specs  d1a83c8ee803c3b5edf4786a47a55da0f3e16617
Remove microversions spec from backlog

This was a discussion that we had at the Pike PTG in Atlanta and is
recapped in our weekly meeting:

  http://eavesdrop.openstack.org/meetings/keystone/2017/keystone.2017-02-28-18.00.log.html#l-47

This commit removes the microversions spec from the backlog because
there are some overall decisions regarding microversions that need to
be made at the API-WG level. Once those issues are resolved, we can
repropose a more accurate representation of the spec.

Change-Id: Ie4185344a150020785071d914c25cb901ba89b3a