Updated openstack/openstack Project: openstack/tempest  d184d619d7adc6ab60055884515055a7f3cf7770
Remove unnecessary show_server in TestServerAdvancedOps

In test_server_sequence_suspend_resume, show_server follows
wait_for_server_status which is meaningless, because in
wait_for_server_status if server didn't reach the expected
state, an exception will be raised.

This is to remove the unnecessary show_server and rearrange
the code so it can look clearer.

Change-Id: Ifbb3403cc68cf0ea4bd1f84e5972c74598c8d329