Updated openstack/openstack Project: openstack-infra/nodepool  d616e61723207ed0e29ea69d67908a00ebf2cdfb
Add destructor to SSHClient

Newer versions of paramiko require a client object to be explicitly
closed.  Fortunately, we wrap all of our use of paramiko client
objects in our own class.  Add a destructor to our class which
closes the client object.

Note, this has been tested to work (and is needed) even if a
connection is not established.

Change-Id: I5dff7ed254567968b42d053b85004769f8647ecb