Updated openstack/openstack Project: openstack/tripleo-heat-templates  de1637bdd1f7753f8a20469697f98ef3ddaf4e2b
Switch to dict format for docker_puppet_tasks

This change gives the option of docker-puppet.py data to be in a dict
as well as a list. This allows docker_puppet_tasks data to use the
same keys as the top level puppet config data.

If the yaql fu can be worked out to build the top level data,
docker-puppet.py can later drop the list format entirely.

Change-Id: I7e2294c6c898d2340421c93516296ccf120aa6d2