Updated openstack/openstack Project: openstack-infra/shade  708049351fcf13634d84e05e97eae3b0a5f0972c
Rename ClusterTemplate in OpenStackCloud docs

There is no ClusterTemplate object in shade. A
cluster template is a resource in shade as any
other, such as a project or a server, which is
returned as a dict.

This patch renames "ClusterTemplate" to "cluster
template" in order to be consistent with how shade
refers to the rest of the resources in the
OpenStackCloud docs.

In addition, it is not useful for shade users to
know ClusterTemplate is the new name for BayModel.
If shade had ever supported a BayModel resource,
it would still support it.

Change-Id: Icbf4b310c5bbfb912c902f66974e11d8cefe1e23